reconstruct the atomictic library
