+++ mlcolvar/featurization/atomistic/backbones/__init__.py
@@ -1,10 +1,12 @@
+from .base import BaseAtomisticBackbone
 from .deepmd import DeepMDBackbone
 from .mace import MACEBackbone
 from .pet import PETBackbone
 
 
 __all__ = [
-    "DeepMDBackbone",
+    "BaseAtomisticBackbone",
     "MACEBackbone",
     "PETBackbone",
+    "DeepMDBackbone",
 ]
