+++ mlcolvar/featurization/atomistic/backbones/pet.py
@@ -1,18 +1,21 @@
 from __future__ import annotations
 
+# Apply the PET/TorchScript compatibility patch before importing metatensor or metatomic.
+from ..patches import pet_jit as _pet_jit  # noqa: F401
+
 from typing import Dict, List, Optional, Tuple
 
 import torch
 from torch import nn
 
-from mlcolvar.integrations.atomistic import BaseAtomisticBackbone
-from mlcolvar.integrations.utils import (
+from .base import BaseAtomisticBackbone
+from ..graph import (
     get_graph_ptr,
     prepare_cells,
     prepare_pbc,
 )
 
-from ._utils import (
+from .utils import (
     to_bool,
     to_float,
     to_int,
