+++ mlcolvar/featurization/atomistic/backbones/utils.py
