+++ mlcolvar/featurization/atomistic/metatomic/__init__.py
@@ -1,0 +1,16 @@
+from .export import (
+    create_metatomic_model,
+    export_metatomic_model,
+)
+from .wrapper import (
+    CVInferenceModel,
+    MetatomicCVWrapper,
+)
+
+
+__all__ = [
+    "CVInferenceModel",
+    "MetatomicCVWrapper",
+    "create_metatomic_model",
+    "export_metatomic_model",
+]
