+++ mlcolvar/featurization/atomistic/patches/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility patches for optional atomistic dependencies."""
